refactor(fspy-shm): consolidate to three files
The six-way split was designed around machinery the simplifications have
since deleted, and its narrative had drifted: writer and reader derive
their own payload references, so state was never the only file touching
shared memory. Merge to the boundary that still earns its keep — pure
integer math versus code that touches the mapping:

- layout.rs absorbs the descriptor codec (both plain arithmetic)
- shared.rs is state + writer + reader in reading order, with the
  reservation types and SharedState going private to it
- mod.rs keeps the surface, overview docs, and integration tests
